+++ spec/ruby3_backward_compatibility/compatibility/file_spec.rb
@@ -1,0 +1,12 @@
+require 'ruby3_backward_compatibility/compatibility/file'
+
+module Ruby3BackwardCompatibility
+  describe File do
+    describe '#exists?' do
+      it 'works like "exist?"' do
+        expect(File.exists?(__FILE__)).to eq(true)
+        expect(File.exists?('foobar.baz')).to eq(false)
+      end
+    end
+  end
+end