+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Added back `File.exists?` and `Dir.exists?` on Ruby 3.2+.
+
 ## [1.1.1] - 2022-12-01
 
 - Gem no longer requires Ruby 3. It will not do anything when used on Ruby 2.x, but you can add it to your Gemfile if you need that for a Ruby migration strategy. (Thanks to @BigAirJosh)